Simple Search + Cart

+++ src/main/java/ashteam/farm_leftover/product/dto/NewProductDto.java
@@ -2,10 +2,12 @@ package ashteam.farm_leftover.product.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class NewProductDto {
     String productName;
-    Double pricePerUnit;
+    BigDecimal pricePerUnit;
     String unit;
     Integer availableQuantity;
 }
