+++ src/main/java/ashteam/farm_leftover/product/dto/ProductForFarmDto.java
@@ -2,13 +2,14 @@ package ashteam.farm_leftover.product.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
 public class ProductForFarmDto {
     Long productId;
     String productName;
-    Double pricePerUnit;
+    BigDecimal pricePerUnit;
     String unit;
     Integer availableQuantity;
     LocalDateTime createdAt;
